LPD-25046 Move ExpectedLogMethodTestRule to test/rule
