+++ modules/apps/portal-search/portal-search-test-util/bnd.bnd
@@ -2,6 +2,7 @@ Bundle-Name: Liferay Portal Search Test Utilities
 Bundle-SymbolicName: com.liferay.portal.search.test.util
 Bundle-Version: 10.2.5
 Export-Package:\
+	com.liferay.portal.search.test.rule.logging,\
 	com.liferay.portal.search.test.util,\
 	com.liferay.portal.search.test.util.logging
 Import-Package:\
