+++ modules/apps/portal-search/portal-search-test-util/src/main/resources/com/liferay/portal/search/test/rule/logging/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
