+++ portal-test/src/com/liferay/portal/kernel/test/rule/AggregateTestRule.java
@@ -66,7 +66,7 @@ public class AggregateTestRule implements TestRule {
 			"WebDAVEnvironmentConfigClassTestRule",
 		"com.liferay.portal.test.rule.PermissionCheckerMethodTestRule",
 		InitializeKernelUtilTestRule.class.getName(),
-		"com.liferay.portal.search.test.util.logging.ExpectedLogMethodTestRule",
+		"com.liferay.portal.search.test.rule.logging.ExpectedLogMethodTestRule",
 		"com.liferay.portal.security.script.management.test.rule." +
 			"ScriptManagementConfigurationTestRule"
 	};
